requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.13.0
django-filter==21.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
jmespath==0.10.0
numpy==1.22.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
